tests/wpt/tests/element-timing/fixed-id-identifier.html

<!DOCTYPE HTML>
<meta charset=utf-8>
<title>Element Timing: entry does not change its id or identifier value</title>
<body>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/element-timing-helpers.js"></script>
<p elementtiming='my_identifier' id='my_id'>Text</p>
<script>
  async_test(function (t) {
    assert_implements(window.PerformanceElementTiming, "PerformanceElementTiming is not implemented");
    const observer = new PerformanceObserver(
      t.step_func_done(function(entryList) {
        assert_equals(entryList.getEntries().length, 1);
        const entry = entryList.getEntries()[0];
        assert_equals(entry.id, 'my_id');
        assert_equals(entry.identifier, 'my_identifier');
        const element = document.getElementById('my_id');
        element.id = 'other_id';
        element.setAttribute('elementtiming', 'other_identifier');
        assert_equals(entry.id, 'my_id');
        assert_equals(entry.identifier, 'my_identifier');
      })
    );
    observer.observe({type: 'element', buffered: true});
  }, 'PerformanceElementTiming id and identifier do not change when Element changes.');
</script>
</body>
